Builds and passes all unit tests without fail
Although the doctests have changed, there are no changes to the
semantics of the tests (except for the deprecation of shapely/wkb.loads)
Any changes made were simply so that the python3 interpreter would run
the tests

For better py3k support, wkb.loads and wkb.dumps have been deprecated
in favour of wkb.load_bytes and wkb.dump_bytes. The original functions
will be added back in the next commit with an appropriate deprecation
warning.

The reasoning for this is that the wkb is properly a byte format, so
they shouldn't have been str to begin with. The wkt format are ascii
strings, so the loads and dumps functions have been preserved in this
case